source/requirements.txt
gitpython
requests
nc_py_api
python-dotenv
PyGithub
pypandoc